fix(checks): controller addendum -- markdownlint on SHARED_CHECKS.md, pyright on orchestrator.py
Two real findings the controller's clean-worktree store run surfaced at
57704b70, fixed at the source:

(a) hook.markdownlint-cli2 FAIL on docs/SHARED_CHECKS.md: two lines over
the 120-column cap and one fenced code block missing a language tag
(the C7g docs section). Wrapped the prose, tagged the block ```text.
Store markdownlint-cli2 now reports 0 errors on the file.

(b) types.python FAIL on configs/claude/scripts/agents/orchestrator.py:
`_bounded_probe_output`'s `timeout` parameter was typed `int` while its
only caller-side value (`probe_timeout: float = 15`) and its only use
(`asyncio.wait_for(timeout=timeout)`, which accepts float) are both
float. Widened the parameter to `float` -- the honest type for what the
function actually does -- rather than narrowing the caller or ignoring
the finding. Store pyright on the whole file: 3 errors -> 0.

+++ configs/claude/scripts/agents/orchestrator.py
@@ -119,7 +119,7 @@ if HAS_ANTHROPIC:
     from anthropic import AsyncAnthropic
 
 
-async def _bounded_probe_output(proc, timeout: int, provider: str):
+async def _bounded_probe_output(proc, timeout: float, provider: str):
     """Wait for a provider probe while draining both streams within fixed caps."""
     stdout_task = asyncio.create_task(_read_bounded_stream(proc.stdout))
     stderr_task = asyncio.create_task(_read_bounded_stream(proc.stderr))
